pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "slide-smith"
version = "2.0.0"
description = "Deterministic PowerPoint creation and editing tool"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Bo Zhang" }
]
dependencies = [
  "python-pptx>=1.0.0",
  "pytest>=9.0.0",
  "jsonschema>=4.22.0",
]

[project.scripts]
slide-smith = "slide_smith.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
slide_smith = ["schemas/*.json"]
